+++ game_survivors/src/modules/system_spawner/spawner_components.hpp
@@ -13,7 +13,7 @@ struct EnemySpawnWave
 {
   int wave_start_seconds = 0;
   int wave_end_seconds = 20 * 60;
-  int hp = 30;
+  float hp = 30;
   int max_allowed = 2;
   int number_per_spawn = 1;
   int spawn_cooldown = 4;
